changed to float for balance

+++ question4.py
@@ -50,12 +50,12 @@ def main():
 		if n[3] == "R":
 			acc, name, service, minutes = n[0],n[1],n[3],int(n[4])
 			charge = computeRegular(minutes)
-			balance = int(n[2]) + charge
+			balance = float(n[2]) + charge
 			charges.append(acc,name,service,charge,balance)
 		else:
 			acc, name, service, daytime,offpeak = n[0],n[1],n[3],int(n[4]),int(n[5])
 			charge = computePremium(daytime,offpeak)
-			balance = int(n[2]) + charge
+			balance = float(n[2]) + charge
 			charges.append(acc,name,service,charge,balance)
 	print(charges)
